Moved Vec3 and Orientation to util folder.
