+++ python_example/util/vec.py
